requirements.txt
# VTK must be installed via conda-forge build.
# "pip install vtk" installs vtkOSOpenGLRenderWindow (headless-only) — no display output.
#
# conda create -n meshViewer python=3.10
# conda activate meshViewer
# conda install -c conda-forge vtk
# pip install -r requirements.txt

pyvista>=0.47.0
numpy>=1.24.0
psutil>=5.9.0
trimesh>=4.0.0
pillow>=10.0.0
alive-progress>=3.0.0
librosa>=0.10.0
soundfile>=0.12.0
